% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_findProbeDepth.R
\name{findProbeDepth}
\alias{findProbeDepth}
\title{Probe depth}
\usage{
findProbeDepth(file_name_substrings)
}
\arguments{
\item{file_name_substrings}{character vector, containing substrings of file name}
}
\value{
the probe depth in m (string).
}
\description{
Find probe depth based on file name according to a 5-digit system describing depth in cm
}
\examples{
file_name_substrings=c( "files/20131213","BIM13","00001");findProbeDepth(file_name_substrings)
}
