% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_findSep.R
\name{findSep}
\alias{findSep}
\title{Line separator}
\usage{
findSep(file_path)
}
\arguments{
\item{file_path}{string, full path and name of data file}
}
\value{
the line separator (character).
}
\description{
Find line separator character in a data file
}
\examples{
file_path=file.path("Scripts", "extdata", "file_example1.txt");findSep(file_path)
}
