% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_findSiteName.R
\name{findSiteName}
\alias{findSiteName}
\title{Site name}
\usage{
findSiteName(file_name_substrings)
}
\arguments{
\item{file_name_substrings}{character vector, containing substrings of file name}
}
\value{
the site name (string).
}
\description{
Find site name based on file name according to the naming system "3 letters followed by 2 digits"
}
\examples{
file_name_substrings=c( "files/20131213","BIM13","00001");findSiteName(file_name_substrings)
}
