% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flaq-package.R
\docType{package}
\name{flaq-package}
\alias{flaq}
\alias{flaq-package}
\title{flaq: Functions for L'Assistant de Qualification}
\description{
The package flaq contains functions for the application L'Assistant de Qualification created at Pole R&D ECLA (Aix-en-Provence, France). The goal of the app is to read and qualify high-frequency data.
}
\author{
\strong{Maintainer}: Rosalie Bruel \email{rosalie.bruel@ofb.gouv.fr} (\href{https://orcid.org/0000-0001-8351-7539}{ORCID})

Authors:
\itemize{
  \item Chloé Le Bescond \email{chloe.le-bescond@inrae.fr}
}

}
\keyword{internal}
